ADDNATRL

+++ DSA Learning Series/Easy Problems to Get Started/Add_Natural_Numbers.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    int n, sum = 0;
+    long long n, sum = 0;
     cin>>n;
 
     for(int i=1;i<=n;i++){
